pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "openai>=1.99.6",
    "ipython==8.24.0",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "langchain-text-splitters",
    "langsmith",
    "python-dotenv",
    "yfinance",
    "crewai[tools]",
    "crewai-tools",
    "pydantic",
    "firecrawl-py>=4.3.5",
]
